exoplanet_api/requirements.txt
asgiref==3.9.2
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-yasg==1.21.11
gunicorn==23.0.0
inflection==0.5.1
joblib==1.5.2
numpy==2.3.3
nvidia-nccl-cu12==2.28.3
packaging==25.0
pandas==2.3.3
pillow==11.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
scikit-learn==1.5.1
scipy==1.16.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
uritemplate==4.2.0
xgboost==3.0.5
